feat: Finally fixed the DB bug related to seeding and changed meteorological schema, model, crud, and seed files to all have proper and same column names to solve bug.

+++ migrations/versions/d63493af0a71_create_meteorological_wave_data.py
@@ -26,7 +26,7 @@ def upgrade():
         sa.Column("wind_speed", sa.Float(), nullable=True),
         sa.Column("wind_gust", sa.Float(), nullable=True),
         sa.Column("wave_height", sa.Float(), nullable=True),
-        sa.Column("dominant_wave_period", sa.Integer(), nullable=True),
+        sa.Column("dominant_wave_period", sa.Float(), nullable=True),
         sa.Column("average_wave_period", sa.Float(), nullable=True),
         sa.Column("wave_direction", sa.String(), nullable=True),
         sa.Column("sea_level_pressure", sa.Float(), nullable=True),
